feat: support function fetchmany() and fetchall() in python SDK and UT (#1215)
Reference: [Standard DBAPI](https://www.python.org/dev/peps/pep-0249/#fetchone)
*.fetchone()
Fetch the next row of a query result set, returning a single sequence, None when no more data is available.

*.fetchmany(size=None)
Fetch the next set of rows of a query result, returning a sequence of sequences (e.g. a list of tuples). An empty sequence is returned when no more rows are available.
The number of rows to fetch per call is specified by the parameter. If it is not given, the cursor's arraysize determines the number of rows to be fetched. 

*.fetchall()
Fetch all (remaining) rows of a query result, returning them as a sequence of sequences (e.g. a list of tuples).