feat(udf): hash64 function (#2754)
The function is called `hash64`, or `farm_fingerprint` same as BigQuery. `hash` is not used because it is a reserved keyword in parser.

This is a non-cryptographic hash function, so do not use in cryptographic way. 

The underlaying hash algorithm uses googles [farmhash](https://github.com/google/farmhash), also borrowed from BigQuery. Same result is guaranteed to produced as long as both the data type and value is the same, no matter when to execute the query, or where the executing machine is. 

`hash64` accept a single parameter of type bool, int16, int32, int64, float, double, timestamp, string, date.

Know Limitations

-  `hash64(*)` or hash multiple parameters does not supported yet. Maybe left as an enhancement later. 
- `hash(NULL)` produce `NULL` currently, which might not as expected behavior compared to spark sql etc.