fix: test in src & refactor Explain
Now it is preferred to get dependent tables from Physical Plan instead
of logic plan previously. The behavior has changed inside Explain due to
this, since a TablePlan logic plan node may not refer to a actual table,
but also refer to a CTE in a query contains WITH clause.